Extract Student(Room|Seating)DetailCell
